fixup! Add "number" type to the generator

+++ v3/schemas.go
@@ -2233,7 +2233,7 @@ type Organization struct {
 	// Organization address
 	Address string `json:"address,omitempty"`
 	// Organization balance
-	Balance int `json:"balance,omitempty"`
+	Balance float64 `json:"balance,omitempty"`
 	// Organization city
 	City string `json:"city,omitempty"`
 	// Organization country
